+++ problem-solver/cxx/inferenceModule/logic/NegationExpressionNode.hpp
@@ -17,13 +17,7 @@ public:
 
   void compute(LogicFormulaResult & result) const override;
 
-  LogicFormulaResult generate(Replacements & replacements) override
-  {
-    return {false, false, {}};
-  }
+  void generate(Replacements & replacements, LogicFormulaResult & result) override;
 
-  ScAddr getFormula() const override
-  {
-    return {};
-  }
+  ScAddr getFormula() const override;
 };