+++ problem-solver/cxx/inferenceModule/test/units/ConfigGenerators.hpp
@@ -49,4 +49,4 @@ public:
   }
 };
 
-}  // namespace inference::generator
+}  // namespace inference::generatorTest